requirements.txt
flask

firebase-admin
openai

openai-whisper